app-migrate/requirements.txt
Flask==0.10.1
SQLAlchemy==1.1.1
sqlalchemy_migrate==0.10.0
Flask-SQLAlchemy==2.1
MySQL-python==1.2.5
sqlacodegen==1.1.6
graphene-sqlalchemy==1.1.1
Flask-GraphQL==1.4.1